allow more specificity in install.py
install.py options have changed. The user must now indicate the path to
the project on which packages will be installed and the path from which
the packages will come. A list of packages is no longer accepted (for
now) because the only supported use case at this time is installing
iRODS packages on iRODS containers. This may change in the future to
support a more generic solution, but that is not the goal of this
particular module.